Adds artwork provider system and Calibre plugin

+++ src/Lumina.Contracts/Requests/MediaLibrary/WrittenContentLibrary/BookLibrary/Books/AddBookRequest.cs
@@ -39,7 +39,7 @@ public record AddBookRequest(
     WrittenContentMetadataDto? Metadata,
     BookFormat? Format,
     string? Edition,
-    int? VolumeNumber,
+    float? VolumeNumber,
     BookSeriesDto? Series,
     string? ASIN,
     string? GoodreadsId,
